requirements.txt
openai
tiktoken
langchain
unstructured[pdf]
sentence-transformers
pandas
python-dotenv
redisvl>=0.1.0
redis